www: opt blog posts out of ssr rendering
Removing an upstream useSearchParams opted blog posts into SSR, but it
doesn't work because of next-mdx-remote/codehike, so opting them back
out.
